requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
asyncio==3.4.3
blinker==1.9.0
boto3==1.38.37
botocore==1.38.37
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.57.0
debugpy==1.8.14
decorator==5.2.1
docker==7.1.0
executing==2.2.0
fastapi==0.115.13
filelock==3.18.0
flask==3.1.1
fonttools==4.58.1
fsspec==2025.5.1
gitdb==4.0.12
gitpython==3.1.44
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
idna==3.10
importlib-metadata==8.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.8
mako==1.3.10
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mlflow==3.1.0
mlflow-skinny==3.1.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
optuna==4.3.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt-toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.2.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
werkzeug==3.1.3
zipp==3.23.0
